Improve coherentspinstate speed and stability
By making the the implementation of `coherentspinstate` full recursive
instead of exponentiating inside the loop we get a considerable speedup
for larger spin number. I measured about a factor of 2 for S=100 and 6
for S=1000.

The calculation is also more stable: the growing binomial coefficients
are moderated by the decreasing α^n, avoiding an early overflow. Before,
an overflow would occur for S=2000, now we can do at least 10 times more
before we overflow. There's going to be some loss of precision for very
large S, but up to S=10000 or so it should be ok.